virt: Unify Drive size units
Previously we used to keep some Drive size values in bytes and some in
megabytes, and return next and max volume size in megabytes. We had lot
of code converting sizes from megabytes to bytes. This mess makes
the code harder to understand and error prone.

Now we keep all size values in bytes and return next volume size in
bytes, simplifying lot of code.

Change-Id: Ice01c9706fc27d4472c90e47bf7e808564ba0bfe
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/37727
Reviewed-by: Adam Litke <alitke@redhat.com>