network: api: allow removal of bond with no slaves
If we try to remove a bond with no slaves attached, VDSM explodes and
doesn't allow us to repair the network.

This patch moves checking of slaves up the stack, only to 'edit' and
'add' bond's functions.

'remove' bond function is able to objectivize slaveless bond and allows us
to remove it.

Change-Id: I8b472305c00e3f11a4f7ede78f84eea589ca616b
Signed-off-by: Petr Horáček <phoracek@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1183728
Reviewed-on: http://gerrit.ovirt.org/37751
Reviewed-by: Dan Kenigsberg <danken@redhat.com>