tests: disable memoization in testrunner
memoization during tests is harmful, because
it adds another one hidden layer of global state
shared across test runs.

This may lead -and actually did- to hard
to debug failures in apparently unrelated code.

The net result is waste of developer's time,
and decreased trust in automated tests
because of mysterious failures.

To fix that, we now monkeypatch in testrunner
only utils.memoized with a fake which avoids caching.

Change-Id: Ibd986403a76268995b83c9e6ac400e942ed24cb6
Signed-off-by: Francesco Romani <fromani@redhat.com>
Bug-Url: https://bugzilla.redhat.com/1267851
Reviewed-on: http://gerrit.ovirt.org/37748
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Tested-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/39949