ifcfg: convert yes/no <-> True/False using _from_ifcfg_bool and _to_ifcfg_bool
Use them to format the defaultRoute=True/False property to DEFROUTE=yes/no
without the need to replace the value in IpConfig's namedtuple first and
also to override the property with information from an existing ifcfg file.

If DEFROUTE is not present it is stored as None and not printed to an ifcfg
file at all. To simplify the situation (as seen in other configurators) only
True/False could be used in the future not only in defaultRoute but in other
IP configuration parameters.

Change-Id: I87069ee99fb39d4ce100ccc03117c5a7c88b3e96
Signed-off-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/37792
Reviewed-by: Petr Horáček <phoracek@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Petr Horáček <phoracek@redhat.com>