ifcfg: stop reading from NetDevice.ipConfig namedtuple in _getIfaceConfValues
_getIfaceConfValues updates a copy of a NetDevice's IP configuration before
actually configuring it. The changes must not be brought back to the original
IpConfig so deep copy is performed.

The namedtuple is still returned by _getIfaceConfValues (for now) but in
the next patch the copied IpConfig will be returned directly.

ipconfig.bootproto is not compared against None for compatibility reasons
(it could contain a zero-length string, for instance). It will be converted
to a boolean 'dhcpv4' soon.

Change-Id: Ic72f387b45023527d14a17ba1b44b04d620243e2
Signed-off-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/37793
Reviewed-by: Petr Horáček <phoracek@redhat.com>
Reviewed-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>