virt: periodic: port updateVolume
During the lifetime of a VM, its volumes need
to be updated. This is a task carried by the
'updateVolumes' operation, which was formerly
piggybacked in the sampling threads.

But actually 'updateVolumes' is a perfect candidate
to motivy the need of a 'periodic operations' infrastructure
to run bookkeeping tasks of the VMs.

So, this patch bootstraps the usage of the newly
added 'periodic' infrastructure by moving 'updateVolumes'
away from sampling threads, and making it a periodic operation.

Change-Id: Idf5f4bff03acf4fb43eb6d296e4c02a6b7a52baa
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/37594
Reviewed-by: Dan Kenigsberg <danken@redhat.com>