vm: Introduce Vm._getVolumeSize() method
There are too many places calling self.cif.irs.getVolumeSize(), then
checking for nonzero res["status"]["code"], raising random exception,
or converting returned strings to integer.

This patch introduce a new method that:

- Raise StorageUnavailableError on nonzero status code, so we don't have
  to check for errors when we cannot handle them, and we cannot forget.

- Return VolumeSize namedtuple with integer values, making the calling
  code less error prone and nicer to work with.

This change reveal one call site that was not checking for
getVolumeSize() errors, and could fail with KeyError. Now it will fail
with StorageUnavailableError which is more helpful for debugging.

Change-Id: I2767cb8d7560b0eb1ad950f0844a8e24347139e4
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/37888
Reviewed-by: Adam Litke <alitke@redhat.com>