virt: Introduce Vm._setVolumeSize helper
We have two place invoking self.cif.irs.setVolumeSize(), ignoring errors
without documenting the reason. This patch adds Vm._setVolumeSize()
helper that raises on errors, eliminating such code.

We use setVolumeSize() when extending a disk. First we set volume size to
0, marking it as in transaction. Then we update the guest about the new
size via libvirt, and finally we store the new size in the volume.

In the first setVolumeSize() call setting volume size to 0, it is clear
that we want to fail the operation, so using _setVolumeSize() the
request fails without additional error handling code.

The second setVolumeSize() call updating the volume to the actual new
size is done after the vm was updated and can use the new device, so it
is not clear if we should fail the request or just warn about the
incorrect meta data in the volume (size=0). This patches changes the
behavior to fail the request in this case.

Change-Id: I742fde77a23387f89fa746c2030e151237cb230f
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/37943
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1251958
Reviewed-on: https://gerrit.ovirt.org/45949
Tested-by: Adam Litke <alitke@redhat.com>
Continuous-Integration: Francesco Romani <fromani@redhat.com>