monitor: return whether the reported status is actual
When the domain monitoring results are reported after the domain
monitor was started. The first monitoring run may haven't been yet
completed - which causes the returned status to be determined as OK
by the engine (leads to domain status change).

It was attempted to change the inital returned status in the past but it
broke the host activation flow in the engine (see change I8e0df) and
therefore the inital results can't be changed (backward comptability) -
therefore in this change i add a new info that indicates whether the
reported status is actual or not.

When providing the engine the information on whether the returned status
is actual or not, the engine can decide how to act upon the monitoring
result and ignore it if it's irrelevant.

Change-Id: I1fea518991a76ea0f9ff1ff5258afe95bca2f00d
Bug-Url: https://bugzilla.redhat.com/1183977
Signed-off-by: Liron Aravot <laravot@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/37952
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>