tests: Enhance ReadSpeed.testReadSpeedRegExp unit test
Test changed to be more readable.

Currently, the test defines a parameterized template and a list of
outputs to be tested. For each output, the test replaces
the parameters in the template with the corresponding values from
the output.

This patch makes the test more straightforward by removing the
templating mechanism.

Change-Id: Iec1eaa570729f8db2e00b19181b94daeedbce23b
Signed-off-by: Ala Hino <ahino@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/38021
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Tested-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>