virt: Continue to collect extend candidates after errors
Previously, if we had one bad drive that libvirt fail to get watermarks,
we would stop collecting extend candidates and log a traceback. This
would cause a vm to pause, since we may never extend any of its disks,
and the log would be littered with unhelpful tracebacks.

Now we log an error in such case and continue to collect extend
candidates, so error with one disk cannot effect other disks.

Change-Id: I5d321001d096c45f5b78982c3c0f225728bd4184
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/38045
Reviewed-by: Adam Litke <alitke@redhat.com>