virt: support per-vm stats age
It is possible that a VM goes rougue right at the startup,
even before that any commmand is sent to it.
To cover this case, VDSM marks a VM as unresponsive if stats
are unavailable for too long, including the startup.

This patch provides support in StatsCache to provide the
same behaviour when switching to the new bulk stats sampling.

Change-Id: Id8cd75631c328aa380c10f1f6dd1a8075b2fe0ed
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/38066
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>