misc: Support zero-time in getReadDelay()
We found out that dd returned the following output:

    "517 bytes (517 B) copied, 0 s, Infinity B/s"

The root cause for getting zero/Infinity stats in dd is erratic
behavior of clock_gettime. Testing in the field shows that
clock_gettime() using CLOCK_MONOTONIC may return the same value
for duration of up to 11 milliseconds.

We ran a tool sampling current time using clock_gettime(), and
recording the uniqe readings and the difference between the
previous and the current sample:

    #samples  difference
     177      0.000000001
     175      0.000999812
    9462      0.000999813
     183      0.000999814
       1      0.001999626
       1      0.010997944
       1      0.011997756

In dd, if the start time is equal to the end time, dd returns
"0 s, Infinity B/s"

misc.readpeed() fails to parse that output. This causes the domain
monitor to fail and the domain to move into invalid state. This failure
could cause the host to become non-operational, leading to migration
storm while storage is actually valid.

misc.readpeed() changed to parse the returned output.

Engine tested and is not affected by this change.

Change-Id: I58ae0ec5f3f88a1738ef1b95688534c8c8022362
Bug-Url: https://bugzilla.redhat.com/1191549
Signed-off-by: Ala Hino <ahino@redhat.com>
Signed-off-by: Tobias Fiebig <tfiebig@wybt.net>
Reviewed-on: https://gerrit.ovirt.org/38082
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>