vm: Fix LSM when using libvirt >= 1.2.8
Libvirt 1.2.8 introduced a regression where disk type is converted from
"block" to "file" after blockRebase. This breaks our extend logic, leading
to unwanted extension of up to the disk virtual size, practically making
it preallocated. In 3.5, this used to extend the disk without limit
using all free space on the vg.

Libvirt 1.2.9 introduced a new VIR_DOMAIN_BLOCK_REBASE_COPY_DEV flag
for fixing this issue. This flag must be used when calling blockRebase
with block-based disk. This fix was backported to EL7.1 in
libvirt-1.2.8-16.el7_1.1 and libvirt-python-1.2.8-6.el7_1.1.

This patch uses VIR_DOMAIN_BLOCK_REBASE_COPY_DEV when available, fixing
this issue in EL7.1 and Fedora 21. Other platforms are not affected by
this issue because they are running older libvirt versions.

We assume that we are always working with local libvirt daemon, and
having the flag in the libvirt python bindings means that libvirt daemon
support the flag.

Change-Id: Idf8a53db81b61378ca00ffb8bf069ae944b78db7
Bug-Url: https://bugzilla.redhat.com/1176673
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/38142
Reviewed-by: Eric Blake <eric.b.blake@gmail.com>
Reviewed-by: Adam Litke <alitke@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>