virt: periodic: dispatch work to VMs
This patch adds a class to demultiplex an Operation on per-VM
operation. This is useful to minimize interference
among VMs in case of one operation gets stuck.

Alternative approaches have veen considered, but them
did not provided clear advantages to this approach:

- use per-VM operations, like it was done in the earliest
scalable_sampling concepts. This guarantees the highest
VM safety, as no more than one sampling operation can
be done to a given VM in a given time instant, but fits
poorly with libvirt bulk stats API, and add pressure to
Scheduler and Executor (OTOH, we have object living longer).

- move demultiplexing in Operation class this way Operation
becomes less generic, and more tied to VM.

Change-Id: I7d0422bfce99d1e479cb5af49c7b37204ca7e295
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/37885
Reviewed-by: Dan Kenigsberg <danken@redhat.com>