init: fix network service status checking
In init file, we check if network service is running by its status
return code. Problem is, that this code is "always" 0, therefore we never
start network service when needed (for example, it's not autostarted
on CentOS 6).

`service network status` returns lists of configured and active devices,
if the list of active devices (on the last row) does not contains all
VDSM's ONBOOT=yes devices, it means, that network service did not start
our devices and we should call service network start.

Change-Id: Idfe796744e44af21f08974f2ccb4bbb503cc6d67
Signed-off-by: Petr Horáček <phoracek@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/38173
Reviewed-by: Dan Kenigsberg <danken@redhat.com>