json-rpc: fix the Host.getVMList return value
Host.getVMList() should return either VmDefinition or
VmShortStatus depending on the parameter fullStatus.

The return value in case fullStatus=False was wrong,
as it was just a list of UUIDs, while VmShortStatus
is supposed to be a dictionary holding (at least)
vmId and vmStatus.

This change made life harder to Engine monitoring code.
Missing the status, Engine's monitoring had no choice
but to call Vm.getStats() after each getVmList(), and that
has a not-negligible performance cost.

This patch makes the JSON-RPC output compliant to the
schema, to what Engine's monitoring expects and also
to XML-RPC output.

Change-Id: I2b8b2ed205385f4bdd341cdc576b44312c3fc117
Bug-Url: https://bugzilla.redhat.com/1196327
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/38172
Reviewed-by: Omer Frenkel <ofrenkel@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Tested-by: Omer Frenkel <ofrenkel@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>