virt: Allocate additonal space for cow internal data
In change I4098dfc07 (virt: Fix limit when calculating next volume size) we
fixed drive size limit so drive volume is limited by drive capacity. This
change introduced a regression where vm is paused when trying to fill up a
drive, and there is no way to resume it, since qemu cannot complete the write.

To allow a guest to use the entire drive capacity, we must allocate extra space
for cow internal data. Since it is tricky to compute the exact value, we use
the same estimate (1.1) used for converting raw volumes to cow and merging
volumes.

Change-Id: I14ae7647e459e35d6b6aea632ad4ad9a3cf7666e
Relates-To: https://bugzilla.redhat.com/1195768
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/38178
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>