virt: Avoid pointless extension requests
After a drive is extended to the maximum volume size, if the guest
keeps writing, the image may pass the drive's watermark for early
extension.

On this condition, current extend logic is trying to extend the drive
every 2 seconds. This results in fruitless extensions failing in 
lvextend because the requested size is equal to the current size. These
extension requests may delay real requests, perform unneeded io, and fill
up the log with bogus errors and warnings.

Now we check if the drive was fully extended and do not try to extend it
further.

Change-Id: I0665fb12dbba9d9d42777db4c43cc606b9708348
Bug-Url: https://bugzilla.redhat.com/1195768
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/38179
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>