hostdev: fix creation and deletion of multiple udev rules
When attaching multiple devices from the same group, each of the
devices needs to be detached and then later reattached to the host.
The udev rule creation is hooked into this routine to give qemu access to
the /dev/vfio endpoint. This patch corrects behaviour of udev if the file
already exists or does not exist due to creation or deletion of VM.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1099760
Change-Id: I83fffb2f6e42cc1746f30ca2b63c5d594de014e5
Signed-off-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/38208
Reviewed-by: Dan Kenigsberg <danken@redhat.com>