restore_nets: during rollback, ignore removal of a missing bond
If a bond is removed outside vdsm, attempting to restore networks to the
last known-good set of configuration currently fails.

With this patch, if vdsm is asked to remove an already-missing bond
device during an attempt to restore networking, this descripency is
ignored, and restore_nets would go on attempting to restore further
networks, hopefully restoring connectivity to the host.

The accompanying test verifies that restore_nets survive after such
removal. It also verifies that adding a network on top of an
already-existing bond ends with the bonding device being consumed into
the running (and later, persistent) config.

Change-Id: I7b863ffd829026f9ea3030ac77e57cf6521fe6ba
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/38211