gluster: Modified geo-rep verbs to include newly added "user"
In this patch all the geo-replication related verbs have
been modified to include user name as an argument. This
is need to execute all gluster cli commands, when geo-
replication session is set with a non-root user. This
patch also modifies geo-rep list and status verbs to
list the remote user name which was newly added in gluster
cli.

Change-Id: I336e6c8c6d433787bda4b508b87b6bf9a90f137a
Signed-off-by: Darshan N <dnarayan@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/38318
Reviewed-by: Bala.FA <barumuga@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Signed-off-by: Darshan N <dnarayan@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/39933
Continuous-Integration: Bala.FA <barumuga@redhat.com>