qemu-img: Handle image end offset on different lines of output
Current code assumes that qemu-img check output will place the image end
offset in the second line of output but it can appear later:

$ sudo qemu-img check
/rhev/data-center/.../41b2b944-77f0-40d5-a681-c63e0dbb7482
No errors were found on the image.
70088/98304 = 71.30% allocated, 0.17% fragmented, 0.00% compressed clusters
Image end offset: 4594466816

Update the scanner so it can find it on any line from the second until
the end of output.

Change-Id: Id6610baf1702c9075d8512bdb1da0e05527a0060
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1196072
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/38355
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Ala Hino <ahino@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>