virt: Do not count time spent on migrationCreate against timeout
Currently, the measured time for the source migration timeout includes
the time spent on the destination host. This is bad because the time
spent for some actions during the vm creation can be very expensive and
time consuming.
This patch, excludes the time spent during the destination vm creation
from the measured time.

The change is not expected to introduce regressions on the source, since
even now we would hang on the migrationCreate call.

Change-Id: Iafdaa13136757d455e405328173aa3243bea3995
Bug-Url: https://bugzilla.redhat.com/1221597
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/38397
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Yeela Kaplan <ykaplan@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/40919
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>