gluster: Adds new verb to setup geo-rep mount broker.
This patch adds a new verb to setup mount broker for
gluster geo-replication:

geoRepMountBrokerSetup: This verb takes remote user name,
group name and remote volume name as argument. and
sets the mount broker need for non-root geo-replication.

This patch adds a dependency on glusterfs-3.7dev. This
version of glusterfs can be found at:
http://download.gluster.org/pub/gluster/glusterfs/nightly/glusterfs/

Change-Id: I9498c937bdf44b7747253e9bc203cb9eca1e671f
Signed-off-by: Darshan N <dnarayan@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/38444
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>