api: introducing onlyUUID parameter for getVMList verb
Due to vdsm sending only an array of vmids in a response to getVMList
the engine asks separately using VM.getStats which creates unnecessary
load. In order to fix this issue we introduce a new parameter 'onlyUUID'
which is set to True by default. Engine 3.5.0 still wants to get an
array of vmids, but for all other versions we set 'onlyUUID' to False
which returns map of vmids and statuses.


Change-Id: I2d9eb359f6995c09d61a3159e733498fa3b55780
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>
Bug-Url: https://bugzilla.redhat.com/1196735
Related-To: https://gerrit.ovirt.org/#/c/38448
Reviewed-on: https://gerrit.ovirt.org/38449
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Francesco Romani <fromani@redhat.com>