lib: Add vdsm.concurrent library
Move misc.tmap to new module. We will probably need to move some other
utilities to the new module later.  The tmap utility will be used to
simplify fc-scan and multipath-resize helpers.

tmap() was moved as is from misc for easier review. The next patch will
clean it up a little bit. The tests were rewritten in a simpler way.

Change-Id: Iee90f4c201293c3df048cc4ad7123a837d236a38
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/38464
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>