lib: Simplify and generalize concurrent.tmap()
The previous version of tmap had few issues:

- Used non-daemon threads, so the caller application may be blocked on
  shutdown if a mapped function got stuck.
- Assumed that mapped functions do not raise, limiting the usefulness of
  this utility.
- Logged exceptions for functions that do raise - libraries should not
  log behind you back.
- Dropped all errors but the last one, so the caller could not handle
  them.
- Assumed that all errors are equal, and raised the last error.

This version fixes these issues by making tmap() simpler and more
generic. The function is mapped to the values, and the result is a 2
namedtuple (succeeded, value), holding either the result of the
function, or the exception raised by the function.

The caller is responsible now for handling the errors, and logging them
if needed.

Change-Id: I4337db65fe1b326157507b1424a8b4924b91210f
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/38465
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>