network: api: raise error on removal of assigned bond
When we try to remove assigned bond via setupNetworks, we get SUCCESS,
but bond is not removed (we don't allow removal of assigned device).

Now we raise an ERROR when someone tries to remove assigned bond.

This does not apply for delNetwork, where we cannot remove separate
bonding.

Change-Id: I46edad27b27b96a8d04c04f01b2d5f3e465eb417
Signed-off-by: Petr Horáček <phoracek@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/37883
Continuous-Integration: Jenkins CI
Reviewed-by: Ido Barkan <ibarkan@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>