vm: do not poll for scheduler parameters
The scheduler parameters aren't actually stats, but
configuration data[1], which will not change unless
someone asks libvirt to do so.

VDSM is supposed to be the lone controller of libvirt,
so it is safe to assume (as we already do in a quite
lot of flows) that nothing will change behind his back.

So, stop to poll for scheduler parameters, and store
the change instead.

A future patch will add support for tune events, in
the (unlikely) case someone else indeed start
to mess around with libvirt bypassing VDSM.

+++

[1] as we were reminded into
https://bugzilla.redhat.com/show_bug.cgi?id=1199002#c2

Change-Id: I375065b964dd63d19a804e9a9ebd90a2f98153fb
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/38483
Reviewed-by: Martin Sivák <msivak@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>