Cache the result of numaUtils.getVcpuPids
The method that get the current NUMA cpu assignments
burns quite a lot of CPU cycles to parse the libvirt
XML all the time.

This is not necessary as the XML only needs to be
parsed again after it changes.

This patch adds a NUMA vCpuPids cache to improve
the overall NUMA stats performance.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1182094
Change-Id: I135143bccdd29697ec673819212193c0fc217e56
Signed-off-by: Martin Sivak <msivak@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/38564
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>