network: models: tag non-VDSM devices used by VDSM
When we setup network over pre-existing vlaned bond, models.py skip
configuration of pre-existing devices under a vlan (because they are not
to be changed) in order to not ifdown possibly used devices. Due to this
behaviour we don't persist mentioned networks and therefore we are not
able to restore them.

Now we check if device is a part of vdsm, if not, we configure and
persist it.

Change-Id: I4d3ed2b45d618d404bb340100ca07bc57b2f1136
Signed-off-by: Petr Horáček <phoracek@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/38572
Reviewed-by: Dan Kenigsberg <danken@redhat.com>