network: persistence: skip bonds owned and used by vdsm
When upgrading bonds to unified persistence, we should skip foreign bond
(a bond that has ifcfg-bond lacking the "generated by vdsm" header).
However, if any of these bonds is used by a vdsm network, it must be
persisted. Otherwise, calling restore_net removes the bond (when it
cleans current running config) but it fails to re-create it from the
persistent config.

Change-Id: Idb643a0c3681631db7bd875a03ebdb91fd893e70
Signed-off-by: Petr Horáček <phoracek@redhat.com>
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1154399
Reviewed-on: https://gerrit.ovirt.org/38587
Reviewed-by: Dan Kenigsberg <danken@redhat.com>