hsm: Report info needed for refreshing devices
One of the way to increase the size of a Storage Domain
is to increase the LUN size on the storage server.

Devices that were resized on the storage server need to be
resized also on the host.

When GetDeviceList is called, a multipath rescan occurs.
Therefore the underlying paths are refreshed.
However, the PV still have the previous size.
The devices needed to be refreshed can be discovered
by comparing the path capacity and the PV size.

The following fields are added to GetDevicelist:
Capacity of path in the path status section, and size of PV.

See output with the additional fields:

{'GUID': '36001405bab880a35fc349e183390eb8e',
  'capacity': '24696061952',
  'devtype': 'iSCSI',
  'fwrev': '4.0',
  'logicalblocksize': '512',
  'pathlist': [{'connection': '10.35.0.174',
                'initiatorname': 'default',
                'iqn': 'iqn.2015-01.com.rhevuser:444',
                'port': '3260',
                'portal': '1'}],
  'pathstatus': [{'capacity': '24696061952',
                  'lun': '0',
                  'physdev': 'sda',
                  'state': 'active',
                  'type': 'iSCSI'}],
  'physicalblocksize': '512',
  'productID': 'block_backend_l',
  'pvUUID': '07QE9o-UoSl-v8NP-JTkV-nE3F-nI7x-RgsoQM',
  'pvsize': '24293408768',
  'serial': 'SLIO-ORG__bab880a3-5fc3-49e1-8339-0eb8e882dc16',
  'status': 'used',
  'vendorID': 'LIO-ORG',
  'vgUUID': 'hRpSR4-NTBV-z3Ki-K1N6-llGQ-en24-GGVf8Y'}

Relates-To: https://bugzilla.redhat.com/609689
Change-Id: Icfa5cf4321202362f5a00dad7c9a0347ba52ec71
Signed-off-by: Fred Rolland <frolland@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/38675
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>