API: onlyUUID should affect only short status
this patch makes the 'onlyUUID' parameter take effect only
if 'fullStatus' equals False (short status was asked).

depending on 'fullStatus' boolean parameter, the getVMList API can
return either full or short VM status.
Due to a bug originally introduced before 3.5.0, short status
was incorrectly produced as list(UUIDs), but it is meant to be
a list(dict(vmId:UUID, status:vmStatus), )

The 'onlyUUID' parameter was added to preserve
backward compatibility with early Engines 3.5.x, but it is meant
to fix short status (fullStatus=False), and should not always affect
output.

Instead, the current code always produces short and broken status
if 'onlyUUID' equals True, with the side effect that also fullStatus
is affected.

Bug-Url: https://bugzilla.redhat.com/1196327
Related-To: https://bugzilla.redhat.com/1196735
Change-Id: Id05a1bb92100023e8a96ee6f13daa8eedd233ff6
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/38679
Reviewed-by: Dan Kenigsberg <danken@redhat.com>