virt: remove unwanted 'response' aliasing
In quite a few place we used the name 'response'
for variables holding the result of an API call.

After the introduction of the 'response' module,
this created unwanted aliasing, with room for subtle
annoying bugs like

AttributeError: 'dict' object has no attribute 'success'

With some pyflakes, this may also cause the error

vm.py:1234: redefinition of unused 'response' from line 41

thus breaking the build.

To fix this, we look for variables which used the 'response'
name, and we just rename them to 'result', without changing
the program flow.

Change-Id: I53f91af7e6d28aaf033bfb589216c0422bd4a311
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/38803
Reviewed-by: Sandro Bonazzola <sbonazzo@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>