storage: Stop using tempnam in deleteImage
tempnam does not guarantee that the entire requested prefix is part of
the resulting file name but we rely on the whole prefix to be present to
identify the images for garbage collection.  tempnam is a bit of
overkill anyway since we already have a guaranteed unique file name (the
image UUID) and locking to prevent races.

Change-Id: I0ff87e735c12c94d704391286d1bfc451b442956
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/38768
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>