lib: introduce the 'response' module
It happens sometimes that it is needed to overwrite
the predefined error message with something more
specific to the circumstances.

In general, the handling of response code
feels clumsier than it should.

To improve things, we bootstrap the little helper
'response' module, starting with a function
which allow to override the error message.

Change-Id: Ife293080c2c6654315a609478f3276c9008ea6e7
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/37912
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>