API: getVMList: compatibity with internal clients
After the addition of the 'onlyUUID' parameter, the external
API of VDSM is now backward compatible:

- in XMLRPC, we always set explicitely onlyUUID=False, so no room for
  breakage
- in JSONRPC, we have Host_getVMList_Call (short status), on which
  * if onlyUUID is provided from the outside, we use it
    (works with Engine >= 3.5.1)
  * if onlyUUID is NOT given (Engine 3.5.0), we default as True,
    and doing so we preserve the backward compatibility
- in JSONRPC, in Host_getVMFullList_Call (full status), we use
  onlyUUID=False explicitely, so no issue hee

The internal API, however, changed behaviour.
Being the API internal, it doesn't usually need special care,
but this time we need, because MOM is a supported client of
this API.

With the current 'onlyUUID' defaults, clients of API.py
will get a list of UUIDs instead of a list of Short Status.

To fix this, and given the findings outlined in the above bullet points,
we just switch the default value of 'onlyUUID' to cover this use case.

Bug-Url: https://bugzilla.redhat.com/1196327
Related-To: https://bugzilla.redhat.com/1196735
Change-Id: I0560a262329f729716577b28aa78b177f0929121
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/38805
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/38868
Reviewed-by: Martin Sivák <msivak@redhat.com>