hostdev: add after_hostdev_list_by_caps hooking point
In order to allow oVirt deployment maintainers some customization over
which host devices are hidden to oVirt or do modification/quirks for
specific hardware, new hooking point, after_hostdev_list_by_caps is
introduced. This hooking point is another _json_hook, meaning that
it receives device json representation and is free to modify it.

This hook needs to be used with caution: modifying the device forest
(multiple trees) may lead to engine's inability to parse it and use
it correctly. Safe modification are removing leaf nodes and whole
trees. Although engine may not be able to parse broken tree correctly,
it should not explode or cause inconsistent database state.

Change-Id: I4ab6b3fee49cea83266c77dbfcd8c08ed69c630e
Signed-off-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/38884
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>