virt: make _initVmStats public
A followup patch want to add an explicit stopVmStats method.
Currently, we have a private helper to start the Vm Stats Thread:
_initVmStats.

It could be surprising to have a private method to start and
a public one to stop, so this patch renames _initVmStats into
startVmStats to make the flows symmetric, with no further
changes in the code flow.

Change-Id: Id19ed15cc6743dba3baee13643ac39a5814b8ade
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/39036
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>