clientIF: stop VmStatsThread on shutdown
Explicitily stop VmStatsThreads when VDSM is shutting down.
This patch makes the flow faster and safer when libvirtd crashes
Without this patch, the VmStatsThread exit on shutdown because
they find Vm._dom equals None, and they exit with a stacktrace
as soon as they woke up for their polling.

With this patch, we make things explicit.

Relates-To: https://bugzilla.redhat.com/1203210
Change-Id: Id1c01f7a3ea1c22ba0d64b39fb74393e9a52e236
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/39038
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>