schema: temporarily skip getMigrationStatus
Commit b8df0ff261c unveiled an existing inconsistency
betwen the return value of VM.getMigrationStatus() API
and the schema.
As result, schema validation tests, thus master,
are broken.

Investigation on Engine side suggests that Engine
does not really cares about new keys, so we
have the chance to properly fix, cleanup and modernize
this verb with no hurry.

However, to quickly unbreak master, this patch
just skips the schema validation of the getMigrationStatus
verb, until it is fully worked on.

Change-Id: I86b15d626314e679b7fecb6d40cb9d950086fe27
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/39062
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>